feat: cr shouldActivateDisabledStrategies
Refactored FeatureToggleSwitch.tsx to have all logic needed
